% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{input_provided}
\alias{input_provided}
\title{Determine whether a field has a value that can be validated}
\usage{
input_provided(val)
}
\arguments{
\item{val}{Values to test for availability in a Shiny context.}
}
\value{
A logical vector of length 1.
}
\description{
This function is based on \code{\link[shiny:req]{shiny::isTruthy()}} but tweaked here in
\strong{shinyvalidate} to better embody the idea that a value should be available
(i.e., not \code{NULL}) and it should be in a form that is reasonable to validate.
}
