% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{skip_validation}
\alias{skip_validation}
\title{Skip any normal validation performed by a rule}
\usage{
skip_validation()
}
\value{
A function that returns a sentinel value, signaling to shinyvalidate
that any further validation rules for an input are to be skipped.
}
\description{
While the predominant role of the \code{skip_validation()} function is tied to the
\code{\link[=sv_optional]{sv_optional()}} function (where it's used internally), you can also return
\code{skip_validation()} from custom validation rules. When returned, all
subsequent validation rules defined for the input will be skipped.
}
