% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{sv_has_value}
\alias{sv_has_value}
\title{Determine whether a field has a value that can be validated}
\usage{
sv_has_value(x)
}
\arguments{
\item{x}{An expression to test for availability of values and ability to be
validated.}
}
\value{
A logical vector of length 1.
}
\description{
This function is based on \code{\link[shiny:req]{shiny::isTruthy()}} but tweaked here in
\strong{shinyvalidate} to better embody the idea that a field value should be
available (i.e., not \code{NULL}) and it should be in a form that is reasonable
to validate.
}
