% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{sv_in_set}
\alias{sv_in_set}
\title{Validate that a field is part of a defined set}
\usage{
sv_in_set(set, allow_nan = FALSE, message = NULL, msg_limit = 3)
}
\arguments{
\item{set}{A vector of elements for which the field value must be a part of
to pass validation.}

\item{allow_nan}{If \code{FALSE} (the default), then the presence of a \code{NaN} value
element will cause validation to fail.}

\item{message}{The validation error message to use if a value fails to match
the rule. By default, this is generic message provided by \strong{shinyvalidate}
but a custom message can be provided here.}

\item{msg_limit}{The limit of \code{set} values to include in the
automatically-generated error message (i.e., when \code{message = NULL}). If the
number of elements provided in \code{set} is greater than \code{msg_limit} then only
the first \verb{<message_limit>} set elements will be echoed along with text
that states how many extra elements are part of the \code{set}.}
}
\value{
A function suitable for using as an
\code{\link[=InputValidator]{InputValidator$add_rule()}} rule.
}
\description{
The \code{sv_in_set()} function checks whether the field value is part of a
specified set of values.
}
\examples{
# Ignore withReactiveDomain(), it's just required to get this example to run
# outside of Shiny
shiny::withReactiveDomain(shiny::MockShinySession$new(), {

  iv <- InputValidator$new()

  iv$add_rule("count", sv_in_set(1:5))
  iv$add_rule("count", ~if (. <= 0) "A positive value is required")

})
}
