% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{sv_required}
\alias{sv_required}
\title{Validate that the field is present}
\usage{
sv_required(message = "Required", test = shiny::isTruthy)
}
\arguments{
\item{message}{The validation error message to be displayed if the test does
not pass.}

\item{test}{A single-argument function, or single-sided formula (using \code{.} to
access the value to test), that returns \code{TRUE} for success and \code{FALSE} for
failure.}
}
\value{
A function suitable for using as an
\code{\link[=InputValidator]{InputValidator$add_rule()}} rule.
}
\description{
Call \code{sv_required()} to generate a validation function that ensures an input
value is present. By default, the definition of "is present" is based on
\code{\link[shiny:req]{shiny::isTruthy()}}, which is the logic used by the \code{\link[shiny:req]{shiny::req()}} function
as well.
}
\examples{
\dontrun{
iv <- InputValidator$new()

# Basic usage: ensure that `input$title` is present, and return a terse
# validation message if not
iv$add_rule("title", sv_required())

# You can easily provide a custom message to display
iv$add_rule("email", sv_required("An email is required"))

# Provide a `test` argument to change the definition of "is present";
# in this example, any non-NULL value will be accepted.
iv$add_rule("choices", sv_required(test = is.null))
}
}
