% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{input_provided}
\alias{input_provided}
\title{Determine whether a field has a value that can be validated}
\usage{
input_provided(val)
}
\arguments{
\item{val}{Values to test for availability in a Shiny context.}
}
\value{
A logical vector of length 1.
}
\description{
This function is based on \code{\link[shiny:req]{shiny::isTruthy()}} but tweaked here in
\strong{shinyvalidate} to better embody the idea that a value should be available
(i.e., not \code{NULL}) and it should be in a form that is reasonable to validate.
}
\details{
\code{input_provided} returns \code{TRUE} for all values except:
\itemize{
\item \code{NULL}
\item \code{""}
\item An empty atomic vector
\item An atomic vector that contains only missing values
\item A character vector that contains only missing and/or \code{""} values
\item An object of class \code{"try-error"}
\item A value that represents an unclicked \code{\link[shiny:actionButton]{shiny::actionButton()}}
}
}
