% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{sv_regex}
\alias{sv_regex}
\title{Validate based on regular expression}
\usage{
sv_regex(
  pattern,
  message,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE,
  invert = FALSE
)
}
\arguments{
\item{pattern}{Character string containing a regular expression (or character
string if \code{fixed = TRUE}) to be tested against. If a character vector of
length 2 or more is supplied, the first element is used with a warning.}

\item{message}{The validation error message to use if a value fails to match
the pattern.}

\item{ignore.case, perl, fixed, useBytes, invert}{Passed through to \code{\link[base:grep]{base::grepl()}}.}
}
\description{
A validation function, suitable for use with \code{InputValidator$add_rule()},
that checks whether input values match the specified regular expression.
}
